+++ src/main/java/com/example/activity_diary/dto/diary/metric/EntryMetricValueCreateDto.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class EntryMetricValueCreateDto {
@@ -15,5 +17,5 @@ public class EntryMetricValueCreateDto {
 
     @NotNull
     @Positive
-    private Integer value;
+    private BigDecimal value;
 }
