+++ src/main/java/com/example/activity_diary/dto/diary/metric/EntryMetricValueResponseDto.java
@@ -3,12 +3,14 @@ package com.example.activity_diary.dto.diary.metric;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class EntryMetricValueResponseDto {
 
     private Long unitId;
     private String unitName;
-    private Integer value;
+    private BigDecimal value;
 }
