+++ src/main/java/com/example/activity_diary/dto/goal/EntryMetricValueGoalDto.java
@@ -2,6 +2,8 @@ package com.example.activity_diary.dto.goal;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class EntryMetricValueGoalDto {
     Long id;
@@ -9,5 +11,5 @@ public class EntryMetricValueGoalDto {
     Long unitId;
     String unitName; // если надо
 
-    Integer expectedValue;
+    BigDecimal expectedValue;
 }
